Support ducklake database in primary database and add is_ducklake flag to identify motherduck managed ducklake (#625)
Implements the approach mentioned in https://github.com/duckdb/dbt-duckdb/issues/619 to support motherduck-managed ducklake database, which cannot currently be detected because it does not use the `ducklake:` prefix. This is necessary because otherwise dbt treats the managed ducklake database as a regular database then we get ` Not implemented Error: Cascade Drop not supported in DuckLake`.

Also make it possible to have a primary database that is ducklake (both regular `ducklake:...` and motherduck-managed ducklake db), because it is surprising that it didn't work. But since there's no aliasing, the database name will have to be inferred from the path and may not work for all paths. 

## Testing plan
- [x] Unit tests
- [x] ran some example dbt projects with local and motherduck ducklakes. Here are some profiles I tested. When there are additional attached databases, some of the models are routed to those databases.

```
    ducklake_md:
      type: duckdb
      schema: main
      path: 'md:my_ducklake'
      threads: 1
      is_ducklake: true
      disable_transactions: false
      extensions:
        - httpfs

    ducklake_local:
      type: duckdb
      schema: main
      path: 'ducklake:my_ducklake.ddb'
      threads: 1
      disable_transactions: false
      extensions:
        - httpfs
    
    mixed_md:
      type: duckdb
      path: 'md:my_ducklake3'
      is_ducklake: true
      schema: test
      threads: 1
      attach:
        - path: "md:my_not_ducklake"
        - path: "md:my_other_ducklake"
          is_ducklake: true
    
    mixed_local:
      type: duckdb
      path: 'ducklake:my_ducklake3.ddb'
      schema: test
      threads: 1
      attach:
        - path: "my_not_ducklake.ddb"
        - path: "ducklake:my_other_ducklake.ddb"
 ```

The 3rd one needed this fix https://github.com/duckdb/dbt-duckdb/pull/626